+++ Carbon.Accounts/Functions/ConvertTo-CSecurityIdentifier.ps1
@@ -13,10 +13,10 @@ function ConvertTo-CSecurityIdentifier
     If the string or byte array don't represent a SID, an error is written and nothing is returned.
 
     .LINK
-    Resolve-CIdentity
+    Resolve-CPrincipal
 
     .LINK
-    Resolve-CIdentityName
+    Resolve-CPrincipalName
 
     .EXAMPLE
     ConvertTo-CSecurityIdentifier -SID 'S-1-5-21-2678556459-1010642102-471947008-1017'