+++ Carbon.Accounts/Functions/Resolve-CPrincipal.ps1
@@ -1,14 +1,14 @@
 
-function Resolve-CIdentity
+function Resolve-CPrincipal
 {
     <#
     .SYNOPSIS
     Gets domain, name, type, and SID information about a user or group.
 
     .DESCRIPTION
-    The `Resolve-CIdentity` function takes an identity name or security identifier (SID) and gets its canonical
-    representation. It returns a `Carbon_Accounts_Identity` object, which contains the following information about the
-    identity:
+    The `Resolve-CPrincipal` function takes a principal name or security identifier (SID) and gets its canonical
+    representation. It returns a `Carbon_Accounts_Principal` object, which contains the following information about the
+    principal:
 
      * Domain - the domain the user was found in
      * FullName - the users full name, e.g. Domain\Name
@@ -29,10 +29,10 @@ function Resolve-CIdentity
     returned.
 
     .LINK
-    Test-CIdentity
+    Test-CPrincipal
 
     .LINK
-    Resolve-CIdentityName
+    Resolve-CPrincipalName
 
     .LINK
     http://msdn.microsoft.com/en-us/library/system.security.principal.securityidentifier.aspx
@@ -44,41 +44,41 @@ function Resolve-CIdentity
     ConvertTo-CSecurityIdentifier
 
     .LINK
-    Resolve-CIdentityName
+    Resolve-CPrincipalName
 
     .LINK
-    Test-CIdentity
+    Test-CPrincipal
 
     .OUTPUTS
-    Carbon_Accounts_Identity.
+    Carbon_Accounts_Principal.
 
     .EXAMPLE
-    Resolve-CIdentity -Name 'Administrators'
+    Resolve-CPrincipal -Name 'Administrators'
 
     Returns an object representing the `Administrators` group.
 
     .EXAMPLE
-    Resolve-CIdentity -SID 'S-1-5-21-2678556459-1010642102-471947008-1017'
+    Resolve-CPrincipal -SID 'S-1-5-21-2678556459-1010642102-471947008-1017'
 
-    Demonstrates how to use a SID in SDDL form to convert a SID into an identity.
+    Demonstrates how to use a SID in SDDL form to convert a SID into an principal.
 
     .EXAMPLE
-    Resolve-CIdentity -SID ([Security.Principal.SecurityIdentifier]::New()'S-1-5-21-2678556459-1010642102-471947008-1017')
+    Resolve-CPrincipal -SID ([Security.Principal.SecurityIdentifier]::New()'S-1-5-21-2678556459-1010642102-471947008-1017')
 
     Demonstrates that you can pass a `SecurityIdentifier` object as the value of the SID parameter.
 
     .EXAMPLE
-    Resolve-CIdentity -SID $sidBytes
+    Resolve-CPrincipal -SID $sidBytes
 
     Demonstrates that you can use a byte array that represents a SID as the value of the `SID` parameter.
     #>
     [CmdletBinding()]
     param(
-        # The name of the identity to return.
+        # The name of the principal to return.
         [Parameter(Mandatory, ParameterSetName='ByName', Position=0)]
         [string] $Name,
 
-        # The SID of the identity to return. Accepts a SID in SDDL form as a `string`, a
+        # The SID of the principal to return. Accepts a SID in SDDL form as a `string`, a
         # `System.Security.Principal.SecurityIdentifier` object, or a SID in binary form as an array of bytes.
         [Parameter(Mandatory , ParameterSetName='BySid')]
         [Object] $SID
@@ -103,20 +103,20 @@ function Resolve-CIdentity
             Write-Error -Message "SID ""${SID}"" not found." -ErrorAction $ErrorActionPreference
             return
         }
-        return [Carbon_Accounts_Identity]::New($account.DomainName, $account.Name, $SID, $account.Use)
+        return [Carbon_Accounts_Principal]::New($account.DomainName, $account.Name, $SID, $account.Use)
     }
 
     if ($Name.StartsWith('.\'))
     {
         $username = $Name.Substring(2)
         $Name = "$([Environment]::MachineName)\${username}"
-        $identity = Resolve-CIdentity -Name $Name
-        if (-not $identity)
+        $principal = Resolve-CPrincipal -Name $Name
+        if (-not $principal)
         {
             $Name = "BUILTIN\${username}"
-            $identity = Resolve-CIdentity -Name $Name
+            $principal = Resolve-CPrincipal -Name $Name
         }
-        return $identity
+        return $principal
     }
 
     if ($Name.Equals("LocalSystem", [StringComparison]::InvariantCultureIgnoreCase))
@@ -127,7 +127,7 @@ function Resolve-CIdentity
     $account = Invoke-AdvapiLookupAccountName -AccountName $Name
     if (-not $account)
     {
-        Write-Error -Message "Identity ""${Name}"" not found." -ErrorAction $ErrorActionPreference
+        Write-Error -Message "Principal ""${Name}"" not found." -ErrorAction $ErrorActionPreference
         return
     }
 
@@ -139,6 +139,6 @@ function Resolve-CIdentity
         $accountName = $domainName
         $domainName = ''
     }
-    return [Carbon_Accounts_Identity]::New($domainName, $accountName, $sid, $account.Use)
+    return [Carbon_Accounts_Principal]::New($domainName, $accountName, $sid, $account.Use)
 
 }